requirements.txt
Flask
pandas
supabase
python-dotenv
